changed cvRNG to avoid BCC32 internal error
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@786 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ include/opencv/cxtypes.h
@@ -297,7 +297,7 @@ typedef uint64 CvRNG;
 
 CV_INLINE CvRNG cvRNG( int64 seed CV_DEFAULT(-1))
 {
-    CvRNG rng = (uint64)(seed ? seed : (int64)-1);
+    CvRNG rng = seed ? (uint64)seed : (uint64)(int64)-1;
     return rng;
 }
